Extract @agent-relay/cloud-sdk from @agent-relay/sdk
Creates a new public workspace package, @agent-relay/cloud-sdk, that
owns the cross-product identity + provisioning concerns previously
embedded in @agent-relay/sdk:

  - Permission types + scope compiler (AgentPermissions, AgentCli,
    compileAgentScopes, resolveAgentPermissions, ...)
  - JWT minting + local JWKS key management (mintAgentToken,
    LocalJwksSigningKey, createLocalJwksKeyPair, importPrivateKeyPem)
  - Relayfile workspace + ACL seeding (seedWorkspace, seedWorkflowAcls,
    createWorkspaceIfNeeded — RelayFileClient wrappers)
  - Relayfile-mount management (ensureRelayfileMount + MountHandle)
  - Permission audit log (PermissionAuditLog,
    getDefaultPermissionAuditPath)

These never belonged in the broker SDK — they're shared across the
agent-relay product family (relayfile, relayauth, relaycast,
relayflows) and any tool that mints scoped tokens or provisions
relayfile workspaces against an Agent Relay account.

After this move:
  @agent-relay/sdk         — broker, PTY, protocol, AgentRelay client
  @agent-relay/cloud-sdk   — cross-product identity + provisioning
  @agent-relay/sdk depends on @agent-relay/cloud-sdk only for the
  AgentCli type (used by cli-resolver and cli-registry).

Consumers updated:
  - relay CLI's `on provision` / `on start` (subpath imports redirected)
  - removed obsolete @agent-relay/sdk/{provisioner,permissions}
    subpath exports and the corresponding workspace path mappings
  - typecheck script builds cloud-sdk between cloud and SDK

provisioner-mount.test.ts (which actually tests provisionWorkflowAgents,
not the SDK) moves into relayflows.
