+++ packages/cli/tsconfig.json
@@ -32,6 +32,8 @@
       "@agent-relay/utils": ["../utils/dist/index.d.ts"],
       "@agent-relay/utils/*": ["../utils/dist/*"],
       "@agent-relay/cloud": ["../cloud/dist/index.d.ts"],
+      "@agent-relay/cloud-sdk": ["../cloud-sdk/dist/index.d.ts"],
+      "@agent-relay/cloud-sdk/*": ["../cloud-sdk/dist/*"],
       "@agent-relay/cloud/*": ["../cloud/dist/*"]
     }
   },
