+++ packages/cloud-sdk/src/__tests__/token-factory.test.ts
