+++ packages/cloud-sdk/src/__tests__/token.test.ts
