+++ packages/cloud-sdk/src/index.ts
@@ -1,0 +1,19 @@
+/**
+ * @agent-relay/cloud-sdk — shared cloud SDK for the relay platform.
+ *
+ * Covers the cross-product concerns (identity, permissions, tokens, audit,
+ * relayfile/relayauth provisioning) used by the agent-relay CLI, @relayflows/core,
+ * and any other tool that needs to mint scoped tokens or provision relayfile
+ * workspaces under an Agent Relay account.
+ *
+ * The @agent-relay/sdk package stays focused on the broker, PTY, and protocol;
+ * anything cross-product belongs here.
+ */
+export * from './permissions.js';
+export * from './audit.js';
+export * from './compiler.js';
+export * from './local-jwks.js';
+export * from './mount.js';
+export * from './seeder.js';
+export * from './token.js';
+export * from './types.js';
