+++ packages/sdk/package.json
@@ -117,26 +117,6 @@
       "import": "./dist/slack.js",
       "default": "./dist/slack.js"
     },
-    "./permissions": {
-      "types": "./dist/permissions.d.ts",
-      "import": "./dist/permissions.js",
-      "default": "./dist/permissions.js"
-    },
-    "./provisioner/seeder": {
-      "types": "./dist/provisioner/seeder.d.ts",
-      "import": "./dist/provisioner/seeder.js",
-      "default": "./dist/provisioner/seeder.js"
-    },
-    "./provisioner/local-jwks": {
-      "types": "./dist/provisioner/local-jwks.d.ts",
-      "import": "./dist/provisioner/local-jwks.js",
-      "default": "./dist/provisioner/local-jwks.js"
-    },
-    "./provisioner/token": {
-      "types": "./dist/provisioner/token.d.ts",
-      "import": "./dist/provisioner/token.js",
-      "default": "./dist/provisioner/token.js"
-    }
   },
   "files": [
     "dist",
@@ -171,6 +151,7 @@
   },
   "dependencies": {
     "@agent-relay/cloud": "7.1.1",
+    "@agent-relay/cloud-sdk": "7.1.1",
     "@agent-relay/config": "7.1.1",
     "@agent-relay/github-primitive": "7.1.1",
     "@agent-relay/slack-primitive": "7.1.1",
