+++ packages/sdk/src/index.ts
@@ -44,8 +44,6 @@ export * from './consensus.js';
 export * from './shadow.js';
 export * from './relay-adapter.js';
 export * from './harness.js';
-export * from './permissions.js';
-export * from './provisioner/index.js';
 export * from './spawn-from-env.js';
 export * from './cli-registry.js';
 export * from './cli-resolver.js';
