--- packages/sdk/src/provisioner/index.ts
@@ -1,14 +1,0 @@
-/**
- * Provisioner primitives: scope compilation, token minting, workspace
- * seeding, mount management, audit logging.
- *
- * Workflow-shaped orchestration (`provisionWorkflowAgents`) lives in
- * @relayflows/core and composes these primitives.
- */
-export * from './audit.js';
-export * from './compiler.js';
-export * from './local-jwks.js';
-export * from './mount.js';
-export * from './seeder.js';
-export * from './token.js';
-export * from './types.js';
