+++ packages/sdk/tsconfig.build.json
@@ -8,6 +8,8 @@
       "@agent-relay/config": ["../config/dist/index.d.ts"],
       "@agent-relay/config/*": ["../config/dist/*"],
       "@agent-relay/cloud": ["../cloud/dist/index.d.ts"],
+      "@agent-relay/cloud-sdk": ["../cloud-sdk/dist/index.d.ts"],
+      "@agent-relay/cloud-sdk/*": ["../cloud-sdk/dist/*"],
       "@agent-relay/cloud/*": ["../cloud/dist/*"],
       "@agent-relay/github-primitive": ["../github-primitive/dist/index.d.ts"],
       "@agent-relay/slack-primitive": ["../slack-primitive/dist/index.d.ts"]
