+++ packages/sdk/tsconfig.json
@@ -7,6 +7,8 @@
       "@agent-relay/config": ["../config/src/index.ts"],
       "@agent-relay/config/*": ["../config/src/*"],
       "@agent-relay/cloud": ["../cloud/src/index.ts"],
+      "@agent-relay/cloud-sdk": ["../cloud-sdk/src/index.ts"],
+      "@agent-relay/cloud-sdk/*": ["../cloud-sdk/src/*"],
       "@agent-relay/github-primitive": ["../github-primitive/src/index.ts"],
       "@agent-relay/slack-primitive": ["../slack-primitive/src/index.ts"]
     },
